moved code to and exported HTML from PostList.js
