+++ src/scripts/GiffyGram.js
@@ -1,42 +1,13 @@
-import { getPosts } from "./data/provider.js"
+import { PostList } from "./feed/PostList.js"
 
 export const GiffyGram = () => {
 
-    const posts = getPosts()
 
-    const displayPosts = (post) => {
-
-        //convert timestamp to date
-        const getDate = () => {
-            const date = new Date(post.timestamp)
-            return date.toLocaleDateString('en-US')
-        }
-
-        //convert posts data to visible html
-
-        return `
-            <div class='post__remark'>
-                ${post.title}
-            </div>
-            <div class='post__image'>
-                <img src="${post.imageURL}" alt="" />
-            </div>
-            <div class='post__tagline'>
-                ${post.description}
-            </div>
-            <div class='post__date'>
-                Posted on ${getDate()}
-            </div>
-        `
-    }
-    
-    // Show main main UI
-    
     let postsHTML =`
         <h1>Giffygram</h1>
-        <section class='post'>
+        <section class='postList'>
             ${
-                posts.map(displayPosts).join("")
+                PostList()
             }
         </section>
     `
